+++ SourceryTests/Parsing/FileParserSpec.swift
@@ -49,6 +49,28 @@ class FileParserSpec: QuickSpec {
                                 "}")
                         expect(result).to(equal([expectedType]))
                     }
+                    
+                    it("extracts file annotation block") {
+                        let annotations: [[String: NSObject]] = [
+                            ["fileAnnotation": NSNumber(value: true), "skipEquality": NSNumber(value: true)],
+                            ["fileAnnotation": NSNumber(value: true), "skipEquality": NSNumber(value: true), "extraAnnotation": NSNumber(value: Float(2))],
+                            ["fileAnnotation": NSNumber(value: true)]
+                        ]
+                        let expectedVariables = (1...3)
+                            .map { Variable(name: "property\($0)", typeName: TypeName("Int"), annotations: annotations[$0 - 1]) }
+                        let expectedType = Class(name: "Foo", variables: expectedVariables, annotations: ["fileAnnotation": NSNumber(value: true), "skipEquality": NSNumber(value: true)])
+                        
+                        let result = parse("// sourcery:file: fileAnnotation\n" +
+                            "// sourcery:begin: skipEquality\n\n\n\n" +
+                            "class Foo {\n" +
+                            "  var property1: Int\n\n\n" +
+                            " // sourcery: extraAnnotation = 2\n" +
+                            "  var property2: Int\n\n" +
+                            "  // sourcery:end\n" +
+                            "  var property3: Int\n" +
+                            "}")
+                        expect(result).to(equal([expectedType]))
+                    }
                 }
 
                 context("given struct") {