+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improved parsing of closure types
 - Check if Current Project Version match version in podspec in release task
 - Improved swift templates performance
+- Added `// sourcery:file` annotation for source code
 
 ### Bug fixes